# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ParamEmailCode:
    properties:
      captcha_answer:
        type: string
      captcha_id:
        type: string
      email:
        type: string
    type: object
  models.ParamEmailExist:
    properties:
      email:
        type: string
    type: object
...
  models.ParamPhoneCode:
    properties:
      captcha_answer:
        type: string
      captcha_id:
        type: string
      phone:
        type: string
    type: object
  models.ParamPhoneExist:
    properties:
      phone:
        type: string
    type: object
  models.ParamSignUpUsingEmail:
    properties:
      code:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      password_confirm:
        type: string
    type: object
  models.ParamSignupUsingPhone:
    properties:
      code:
        type: string
      name:
        type: string
      password:
        type: string
      password_confirm:
        type: string
      phone:
        type: string
    type: object
...
  models.Post:
    properties:
      author_id:
        type: integer
      community_id:
        type: integer
      content:
        type: string
      id:
        type: integer
      status:
        type: integer
      title:
        type: string
    type: object
host: localhost:9000
info:
  contact:
    email: 1298453249@qq.com
    name: xiaorui zheng
    url: http://www.swagger.io/support
  description: 这是一个热点论坛项目， 能够根据当下热点来向用户展示论坛帖子
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 热点论坛
  version: "1.0"
paths:
...
  /auth/code/captcha:
    get:
      consumes:
      - application/json
      description: 获取图形验证码
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取图形验证码
      tags:
      - Auth
  /auth/email:
    post:
      consumes:
      - application/json
      description: 1. 查看邮箱是否存在 2. 获取图形验证码 3. 给指定邮箱发送短信
      parameters:
      - description: 查询参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/models.ParamEmailCode'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 给指定邮箱发送验证码
      tags:
      - Auth
...
  /auth/phone:
    post:
      consumes:
      - application/json
      description: 1. 查看手机号码是否存在 2. 获取图形验证码 3. 给指定手机号码发送短信
      parameters:
      - description: 查询参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/models.ParamPhoneCode'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 给指定号码发送验证码
      tags:
      - Auth
  /auth/signup/email:
    post:
      consumes:
      - application/json
      description: 使用邮箱进行注册
      parameters:
      - description: 查询参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/models.ParamSignUpUsingEmail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 使用邮箱进行注册
      tags:
      - Auth
  /auth/signup/email/exist:
    post:
      consumes:
      - application/json
      description: 验证邮箱是否存在
      parameters:
      - description: 查询参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/models.ParamEmailExist'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 验证邮箱是否存在
      tags:
      - Auth
  /auth/signup/phone:
    post:
      consumes:
      - application/json
      description: 使用手机号码进行注册
      parameters:
      - description: 查询参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/models.ParamSignupUsingPhone'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 使用手机号码进行注册
      tags:
      - Auth
  /auth/signup/phone/exist:
    post:
      consumes:
      - application/json
      description: 判断手机号码是否存在
      parameters:
      - description: 查询参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/models.ParamPhoneExist'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 判断手机号码是否存在
      tags:
      - Auth
...
  /post:
    post:
      consumes:
      - application/json
      description: 创建帖子
      parameters:
      - description: 参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.Post'
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 创建帖子
      tags:
      - Post
  /post/{id}:
    delete:
      consumes:
      - application/json
      description: 删除post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 删除post
      tags:
      - Post
    get:
      consumes:
      - application/json
      description: 获取某个帖子的信息
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取某个帖子的信息
      tags:
      - Post
...
  /post/posts3:
    get:
      consumes:
      - application/json
      description: 删除某个社区的信息
      parameters:
      - description: 页面码
        in: query
        name: page
        required: true
        type: string
      - description: 页面大小
        in: query
        name: size
        required: true
        type: string
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 删除某个社区的信息
      tags:
      - Post
...
swagger: "2.0"